split js into a resemblance of modules
