+++ src/constants.js
@@ -1,0 +1,1 @@
+export const MIN_NUMBER_OF_TASKS = 2;
