+++ src/data/filters.js
@@ -1,0 +1,33 @@
+const filterMocks = [
+  {
+    name: `all`,
+    count: 120,
+  },
+  {
+    name: `overdue`,
+    count: 3,
+    checked: true,
+  },
+  {
+    name: `today`,
+    disabled: true,
+  },
+  {
+    name: `favorites`,
+    disabled: true,
+  },
+  {
+    name: `repeating`,
+    count: 23,
+  },
+  {
+    name: `tags`,
+    count: 56,
+  },
+  {
+    name: `archive`,
+    count: 21,
+  },
+];
+
+export default filterMocks;
