+++ src/data/tasks.js
@@ -1,0 +1,46 @@
+const taskMocks = [
+  {
+    color: `black`,
+    type: `repeat`,
+    text: `Lorem Ipsum is simply dummy text of the printing and typesetting industry.`,
+    hashtags: [`best ever task`, `repeat`, `cinema`],
+  },
+  {
+    color: `yellow`,
+    type: `deadline`,
+    text: `Many desktop publishing packages and web page editors now use Lorem Ipsum as their`,
+    hashtags: [`cars`, `babies`, `bathroom`],
+  },
+  {
+    color: `pink`,
+    type: `repeat`,
+    text: `Lorem Ipsum comes from sections 1.10.32 and 1.10.33 of "de Finibus Bonorum et Malorum" (The Extremes of Good and Evil)`,
+    hashtags: [`dogs`, `repos`, `bitbucket`],
+  },
+  {
+    color: `green`,
+    type: `deadline`,
+    text: `All the Lorem Ipsum generators on the Internet tend to repeat predefined chunks as necessary, making this the first true generator on the Internet. `,
+    hashtags: [`lorem`, `ipsum`, `dolor`],
+  },
+  {
+    color: `blue`,
+    type: `repeat`,
+    text: `The standard chunk of Lorem Ipsum used since the 1500s is reproduced below for those interested. Sections 1.10.32 and 1.10.33 from "de Finibus Bonorum et Malorum" by Cicero are also reproduced in their exact original form, accompanied by English versions from the 1914 translation by H. Rackham.`,
+    hashtags: [`slayer`, `metla`, `megadeath`],
+  },
+  {
+    color: `yellow`,
+    type: `deadline`,
+    text: `It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.`,
+    hashtags: [`better`, `best`, `the best`],
+  },
+  {
+    color: `black`,
+    type: `repeat`,
+    text: `If you are going to use a passage of Lorem Ipsum, you need to be sure there isn't anything embarrassing hidden in the middle of text`,
+    hashtags: [`by now`, `i am`, `bored`],
+  },
+];
+
+export default taskMocks;
