+++ src/data/templates.js
@@ -1,0 +1,81 @@
+export const buttons = `<div class="card__control">
+    <button type="button" class="card__btn card__btn--edit">
+      edit
+    </button>
+    <button type="button" class="card__btn card__btn--archive">
+      archive
+    </button>
+    <button
+      type="button"
+      class="card__btn card__btn--favorites card__btn--disabled"
+    >
+      favorites
+    </button>
+    </div>`;
+
+export const colorInputs = `<div class="card__colors-inner">
+    <h3 class="card__colors-title">Color</h3>
+    <div class="card__colors-wrap">
+      <input
+        type="radio"
+        id="color-black-2"
+        class="card__color-input card__color-input--black visually-hidden"
+        name="color"
+        value="black"
+      />
+      <label
+        for="color-black-2"
+        class="card__color card__color--black"
+        >black</label
+      >
+      <input
+        type="radio"
+        id="color-yellow-2"
+        class="card__color-input card__color-input--yellow visually-hidden"
+        name="color"
+        value="yellow"
+      />
+      <label
+        for="color-yellow-2"
+        class="card__color card__color--yellow"
+        >yellow</label
+      >
+      <input
+        type="radio"
+        id="color-blue-2"
+        class="card__color-input card__color-input--blue visually-hidden"
+        name="color"
+        value="blue"
+      />
+      <label
+        for="color-blue-2"
+        class="card__color card__color--blue"
+        >blue</label
+      >
+      <input
+        type="radio"
+        id="color-green-2"
+        class="card__color-input card__color-input--green visually-hidden"
+        name="color"
+        value="green"
+      />
+      <label
+        for="color-green-2"
+        class="card__color card__color--green"
+        >green</label
+      >
+      <input
+        type="radio"
+        id="color-pink-2"
+        class="card__color-input card__color-input--pink visually-hidden"
+        name="color"
+        value="pink"
+        checked
+      />
+      <label
+        for="color-pink-2"
+        class="card__color card__color--pink"
+        >pink</label
+      >
+    </div>
+  </div>`;
